Move GenericWorkerTyping out of synapse.app
